Improve the panels component
Now, when the panel file is included, only the mixin comes in and you
should call it with the checkboxes data (the names and if it's checked
or not). Another important change is that the `data` object is only in
the files directly rendered by the the controller so it's easier to know
where the data is coming from.